requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.39
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
